Completed the Order Component
